+++ api/src/schema/root/delete-game.js
@@ -41,7 +41,7 @@ export default {
         await tx.raw(
           `
           update players
-          set credits = data.credits::integer
+          set credits = data.credits::bigint
           from (values ${Array.from(betValues, () => '(?, ?)').join(
             ', '
           )}) as data (id, credits)
